+++ src/fah/client/Cores.cpp
@@ -47,7 +47,7 @@ const SmartPointer<Core> &Cores::get(const JSON::ValuePtr &data) {
   string url = data->getString("url");
 
   auto it = cores.find(url);
-  if (it == cores.end()) {
+  if (it == cores.end() || it->second->isInvalid()) {
     SmartPointer<Core> core = new Core(app, data->copy());
     it = cores.insert(cores_t::value_type(url, core)).first;
   }